+++ 2DRPGGAME/Scripts/ResourceScripts/Weapon Actions&Skills/Abilities/AbilityScripts/Archery/Basic/HeavyShot.cs
@@ -6,7 +6,7 @@ public partial class HeavyShot : WeaponBaseAction
 {
     public override void DoAction(CombatActor user, CombatActor target)
     {
-       double damage = user.GetWeaponDMGBySlotType(EquipSlot.Ranged);
+       float damage = user.GetWeaponDMGBySlotType(EquipSlot.Ranged);
         target.TakeDamage(damage);
         user.RemoveCostAfterAction();
     }
