+++ 2DRPGGAME/Scripts/ResourceScripts/Weapon Actions&Skills/Abilities/AbilityScripts/Archery/Basic/RangedShot.cs
@@ -5,7 +5,7 @@ public partial class RangedShot : WeaponBaseAction
 {
 	public override void DoAction(CombatActor user, CombatActor target)
 	{
-		double damage = user.GetWeaponDMGBySlotType(EquipSlot.Ranged);
+		float damage = user.GetWeaponDMGBySlotType(EquipSlot.Ranged);
 		target.TakeDamage(damage);
 		user.RemoveCostAfterAction();
 	}
